RESPONSIVE DESIGN
Completed the Donate page, and made it responsive along with the Home page. The website now
looks good on both desktop and phone screens. Modifying for tablet screens is next. I also added
id's to many of the elements so as to manipulate them with the CSS files I created without interfering
with either the Tailwind or other components in the future.

I also added the link for the Inter font from Google fonts. I know it's a bad practice, but this
commit is fully loaded. So many things in it. Do well to go through everything carefully.